refactor: split AssignmentRoutes.swift (1,401 lines → 2 files), per #491
The new-assignment-creation flow (newAssignmentPage,
updateNewAssignmentDraft, saveNewAssignment, publish, plus three
private helpers) was the largest contiguous block in the file at
~830 lines.  Moved to its own extension file alongside the existing
+Editor / +Submissions / +Enrollment splits.

Layout:
  AssignmentRoutes.swift                — boot + list + validatePage +
                                          open/close/delete/reorder/
                                          updateStatus + brightspace +
                                          editPage (572)
  AssignmentRoutes+NewAssignment.swift  — newAssignmentPage,
                                          updateNewAssignmentDraft,
                                          saveNewAssignment, publish,
                                          private draft helpers (842)

No visibility changes needed — all moved methods were already
internal-visible on the struct.

All AssignmentRoutes tests pass.
